+++ classes/core/wpdk-functions.php
@@ -47,9 +47,6 @@ function wpdk_is_bool( $str )
  * @return bool
  *
  * @file       wpdk-functions.php
- *
- * @deprecated Use WPDKFunctions::is_url
- *
  */
 function wpdk_is_url( $url )
 {
@@ -199,6 +196,7 @@ function wpdk_checked( $haystack, $current, $echo = true )
 function wpdk_selected( $haystack, $current, $echo = true )
 {
   _deprecated_function( __FUNCTION__, '1.2.0', 'WPDKHTMLTagSelect::selected()' );
+
   if ( is_array( $haystack ) && in_array( $current, $haystack ) ) {
     $current = $haystack = 1;
   }