+++ classes/core/wpdk-wordpress-plugin.php
@@ -157,26 +157,7 @@ class WPDKWordPressPlugin extends WPDKPlugin {
    * @see self::urlAjax() static method
    */
   public $urlAjax;
-  /**
-   * The Plugin URL more `assets/css/images/`
-   *
-   * @brief Images URL
-   *
-   * @deprecated Since 0.6.3 - Use `imagesURL` instead
-   *
-   * @var string $url_images
-   */
-  public $url_images;
-  /**
-   * The Plugin URL more `assets/js/`
-   *
-   * @brief Javascript URL
-   *
-   * @deprecated Since 0.6.3 - Use `javascriptURL` instead
-   *
-   * @var string $url_javascript
-   */
-  public $url_javascript;
+
   /**
    * Array key value pairs with list of cron jobs
    *
@@ -244,10 +225,6 @@ class WPDKWordPressPlugin extends WPDKPlugin {
     $this->cssURL        = $this->assetsURL . 'css/';
     $this->imagesURL     = $this->cssURL . 'images/';
     $this->javascriptURL = $this->assetsURL . 'js/';
-
-    /* @deprecated Since 0.6.3 */
-    $this->url_images     = $this->cssURL . 'images/';
-    $this->url_javascript = $this->assetsURL . 'js/';
 
     /* Only folder name. */
     $this->folderName = trailingslashit( basename( dirname( $file ) ) );
@@ -337,14 +314,6 @@ class WPDKWordPressPlugin extends WPDKPlugin {
         }
       }
     }
-  }
-
-  /**
-   * @deprecated Since 0.6.3 - User `currentURL()` instead
-   */
-  public static function current_url() {
-    _deprecated_function( __METHOD__, '0.6.3', 'self::currentURL()' );
-    return self::currentURL();
   }
 
   /**
@@ -580,18 +549,19 @@ class WPDKWordPressPlugin extends WPDKPlugin {
    *
    * @return array
    */
-  public function cron_schedules( $schedules ) {
-    $schedules = array(
+  public function cron_schedules( $schedules )
+  {
+    $new_schedules = array(
       'half_hour'   => array(
-        'interval' => 1800,
+        'interval' => HOUR_IN_SECONDS / 2,
         'display'  => __( 'Half hour', WPDK_TEXTDOMAIN )
       ),
       'two_minutes' => array(
-        'interval' => 60 * 2,
+        'interval' => MINUTE_IN_SECONDS * 2,
         'display'  => __( 'Two minutes', WPDK_TEXTDOMAIN )
       ),
     );
-    return $schedules;
+    return array_merge( $schedules, $new_schedules );
   }
 
   /**
@@ -604,8 +574,8 @@ class WPDKWordPressPlugin extends WPDKPlugin {
    * @param string $hook       Action hook, the execution of which will be unscheduled.
    * @param array  $args       Optional. Arguments to pass to the hook's callback function.
    */
-  public function addCronJob( $recurrence, $hook, $args = array() ) {
-
+  public function addCronJob( $recurrence, $hook, $args = array() )
+  {
     /* If this cron jobs is not scheduled then add to the WP list. */
     if ( !wp_next_scheduled( $hook, $args ) ) {
       wp_schedule_event( time(), $recurrence, $hook, $args );
@@ -626,7 +596,8 @@ class WPDKWordPressPlugin extends WPDKPlugin {
    * @param string $hook
    * @param array  $args
    */
-  public function removeCronJob( $hook, $args = array() ) {
+  public function removeCronJob( $hook, $args = array() )
+  {
     if ( !empty( $this->_cronJobs ) ) {
       $key = md5( serialize( $args ) );
 
@@ -648,7 +619,8 @@ class WPDKWordPressPlugin extends WPDKPlugin {
    * @since 1.0.0.b2
    *
    */
-  public function clearCronJobs() {
+  public function clearCronJobs()
+  {
     $this->removeCronJobs();
     $this->_cronJobs = array();
     update_option( $this->_cronJobsOptionName, $this->_cronJobs );
@@ -660,7 +632,8 @@ class WPDKWordPressPlugin extends WPDKPlugin {
    * @brief Remove all cron jobs
    * @since 1.0.0.b2
    */
-  public function removeCronJobs() {
+  public function removeCronJobs()
+  {
     if ( !empty( $this->_cronJobs ) ) {
       foreach ( $this->_cronJobs as $recurrence => $hooks ) {
         foreach ( $hooks as $hook => $value ) {
@@ -673,7 +646,8 @@ class WPDKWordPressPlugin extends WPDKPlugin {
   /**
    * @deprecated since 0.5 - Use `theme()` instead
    */
-  public function frontend() {
+  public function frontend()
+  {
     _deprecated_function( __METHOD__, '0.5', 'theme()' );
 
     $this->theme();
@@ -684,9 +658,10 @@ class WPDKWordPressPlugin extends WPDKPlugin {
    *
    * @brief Theme
    *
-   * @sa admin()
+   * @sa    admin()
    */
-  public function theme() {
+  public function theme()
+  {
     /* To override. */
   }
 
@@ -694,7 +669,8 @@ class WPDKWordPressPlugin extends WPDKPlugin {
   /**
    * @deprecated since 0.5 - Use `configuration()` instead
    */
-  public function init_options() {
+  public function init_options()
+  {
     _deprecated_function( __METHOD__, '0.5', 'configuration()' );
 
     $this->configuration();
@@ -703,7 +679,8 @@ class WPDKWordPressPlugin extends WPDKPlugin {
   /**
    * @deprecated since 0.7.5 - Use `widgets()` instead
    */
-  public function widgets_init() {
+  public function widgets_init()
+  {
     _deprecated_function( __METHOD__, '0.7.5', 'widgets()' );
 
     $this->widgets();
@@ -716,7 +693,8 @@ class WPDKWordPressPlugin extends WPDKPlugin {
    *
    * @since 0.7.5
    */
-  public function widgets() {
+  public function widgets()
+  {
     /* To override. */
   }
 
@@ -884,8 +862,6 @@ class WPDKPlugin {
     $this->file = $file;
 
     if ( !is_null( $this->file ) ) {
-
-      /* @todo Replace this code below with a custom method. */
 
       /* Use WordPress get_plugin_data() function for auto retrive plugin information. */
       if ( !function_exists( 'get_plugin_data' ) ) {
@@ -1189,7 +1165,6 @@ class WPDKWordPressPaths {
   public static function pluginsURL( $path = '', $plugin = '' ) {
     return trailingslashit( plugins_url( $path, $plugin ) );
   }
-
 
 }